scripts/requirements.txt
boto3==1.34.133
botocore==1.34.133
certifi==2024.6.2
charset-normalizer==3.3.2
docker==7.1.0
idna==3.7
jmespath==1.0.1
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
